# Memory optimisation for GetLastActiveSpan