--- a/benchmarks/Sentry.Benchmarks/SpanIdBenchmarks.cs
+++ b/benchmarks/Sentry.Benchmarks/SpanIdBenchmarks.cs
@@ -1,0 +1,12 @@
+using BenchmarkDotNet.Attributes;
+
+namespace Sentry.Benchmarks;
+
+public class SpanIdBenchmarks
+{
+    [Benchmark(Description = "Creates a Span ID")]
+    public void CreateSpanId()
+    {
+        SpanId.Create();
+    }
+}