--- a/samples/Sentry.Samples.Azure.Functions.Worker/SampleHttpTrigger.cs
+++ b/samples/Sentry.Samples.Azure.Functions.Worker/SampleHttpTrigger.cs
@@ -3,7 +3,7 @@ using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Logging;
 
-namespace Sentry.Samples.AzureFunctions.Worker;
+namespace Sentry.Samples.Azure.Functions.Worker;
 
 public class SampleHttpTrigger
 {