--- a/samples/Sentry.Samples.Azure.Functions.Worker/SampleTimerTrigger.cs
+++ b/samples/Sentry.Samples.Azure.Functions.Worker/SampleTimerTrigger.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
 
-namespace Sentry.Samples.AzureFunctions.Worker;
+namespace Sentry.Samples.Azure.Functions.Worker;
 
 public class SampleTimerTrigger
 {