--- a/samples/Sentry.Samples.Azure.Functions.Worker/Sentry.Samples.Azure.Functions.Worker.csproj
+++ b/samples/Sentry.Samples.Azure.Functions.Worker/Sentry.Samples.Azure.Functions.Worker.csproj
@@ -4,25 +4,25 @@
         <AzureFunctionsVersion>v4</AzureFunctionsVersion>
         <OutputType>Exe</OutputType>
     </PropertyGroup>
-  
+
     <ItemGroup>
         <PackageReference Include="Microsoft.Azure.Functions.Worker" Version="1.13.0" />
         <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.0.13" />
         <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Timer" Version="4.2.0" />
         <PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="1.10.0" />
     </ItemGroup>
-  
+
     <ItemGroup>
         <None Update="host.json" CopyToOutputDirectory="PreserveNewest" />
         <None Update="local.settings.json" CopyToOutputDirectory="PreserveNewest" CopyToPublishDirectory="Never" />
     </ItemGroup>
-  
+
     <ItemGroup>
         <Using Include="System.Threading.ExecutionContext" Alias="ExecutionContext" />
     </ItemGroup>
-  
+
     <ItemGroup>
       <!-- TODO: Replace with a package reference -->
-      <ProjectReference Include="..\..\src\Sentry.AzureFunctions.Worker\Sentry.AzureFunctions.Worker.csproj" />
+      <ProjectReference Include="..\..\src\Sentry.Azure.Functions.Worker\Sentry.Azure.Functions.Worker.csproj" />
     </ItemGroup>
 </Project>