--- a/src/Sentry.Azure.Functions.Worker/Sentry.Azure.Functions.Worker.csproj
+++ b/src/Sentry.Azure.Functions.Worker/Sentry.Azure.Functions.Worker.csproj
@@ -17,7 +17,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <InternalsVisibleTo Include="Sentry.AzureFunctions.Worker.Tests" PublicKey="$(SentryPublicKey)" />
+      <InternalsVisibleTo Include="Sentry.Azure.Functions.Worker.Tests" PublicKey="$(SentryPublicKey)" />
     </ItemGroup>
 
 </Project>