--- a/src/Sentry.Azure.Functions.Worker/SentryAzureFunctionsLoggerProvider.cs
+++ b/src/Sentry.Azure.Functions.Worker/SentryAzureFunctionsLoggerProvider.cs
@@ -2,7 +2,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Sentry.Extensions.Logging;
 
-namespace Sentry.AzureFunctions.Worker;
+namespace Sentry.Azure.Functions.Worker;
 
 [ProviderAlias("Sentry")]
 internal class SentryAzureFunctionsLoggerProvider : SentryLoggerProvider