--- a/src/Sentry.Azure.Functions.Worker/SentryAzureFunctionsOptions.cs
+++ b/src/Sentry.Azure.Functions.Worker/SentryAzureFunctionsOptions.cs
@@ -1,6 +1,6 @@
 using Sentry.Extensions.Logging;
 
-namespace Sentry.AzureFunctions.Worker;
+namespace Sentry.Azure.Functions.Worker;
 
 /// <summary>
 /// Sentry Azure Functions integration options