--- a/src/Sentry/Internal/RandomValuesFactory.cs
+++ b/src/Sentry/Internal/RandomValuesFactory.cs
@@ -7,6 +7,10 @@ internal abstract class RandomValuesFactory
     public abstract double NextDouble();
     public abstract void NextBytes(byte[] bytes);
 
+#if !(NETSTANDARD2_0 || NET461)
+    public abstract void NextBytes(Span<byte> bytes);
+#endif
+
     public bool NextBool(double rate) => rate switch
     {
         >= 1 => true,