--- a/src/Sentry/SpanId.cs
+++ b/src/Sentry/SpanId.cs
@@ -1,4 +1,5 @@
 using Sentry.Extensibility;
+using Sentry.Internal;
 
 namespace Sentry;
 
@@ -7,58 +8,81 @@ namespace Sentry;
 /// </summary>
 public readonly struct SpanId : IEquatable<SpanId>, IJsonSerializable
 {
-    private readonly string _value;
+    private static readonly char[] HexChars = { '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f' };
+    private static readonly RandomValuesFactory Random = new SynchronizedRandomValuesFactory();
 
-    private const string EmptyValue = "0000000000000000";
+    private readonly long _value;
+
+    private long GetValue() => _value;
 
     /// <summary>
     /// An empty Sentry span ID.
     /// </summary>
-    public static readonly SpanId Empty = new(EmptyValue);
+    public static readonly SpanId Empty = new(0);
 
     /// <summary>
     /// Creates a new instance of a Sentry span Id.
     /// </summary>
-    public SpanId(string value) => _value = value;
+    public SpanId(string value) => long.TryParse(value, NumberStyles.HexNumber, null, out _value);
 
-    // This method is used to return a string with all zeroes in case
-    // the `_value` is equal to null.
-    // It can be equal to null because this is a struct and always has
-    // a parameterless constructor that evaluates to an instance with
-    // all fields initialized to default values.
-    // Effectively, using this method instead of just referencing `_value`
-    // makes the behavior more consistent, for example:
-    // default(SpanId).ToString() -> "0000000000000000"
-    // default(SpanId) == SpanId.Empty -> true
-    private string GetNormalizedValue() => !string.IsNullOrWhiteSpace(_value)
-        ? _value
-        : EmptyValue;
+    /// <summary>
+    /// Creates a new instance of a Sentry span Id.
+    /// </summary>
+    /// <param name="value"></param>
+    public SpanId(long value) => _value = value;
 
     /// <inheritdoc />
-    public bool Equals(SpanId other) => StringComparer.Ordinal.Equals(
-        GetNormalizedValue(),
-        other.GetNormalizedValue());
+    public bool Equals(SpanId other) => GetValue().Equals(other.GetValue());
 
     /// <inheritdoc />
     public override bool Equals(object? obj) => obj is SpanId other && Equals(other);
 
     /// <inheritdoc />
-    public override int GetHashCode() => StringComparer.Ordinal.GetHashCode(GetNormalizedValue());
+    public override int GetHashCode() => StringComparer.Ordinal.GetHashCode(_value);
 
     /// <inheritdoc />
-    public override string ToString() => GetNormalizedValue();
+    public override string ToString() => _value.ToString("x8");
 
-    // Note: spans are sentry IDs with only 16 characters, rest being truncated.
-    // This is obviously a bad idea as it invalidates GUID's uniqueness properties
-    // (https://devblogs.microsoft.com/oldnewthing/20080627-00/?p=21823)
-    // but all other SDKs do it this way, so we have no choice but to comply.
     /// <summary>
     /// Generates a new Sentry ID.
     /// </summary>
-    public static SpanId Create() => new(Guid.NewGuid().ToString("n")[..16]);
+    public static SpanId Create()
+    {
+#if NETSTANDARD2_0 || NET461
+        byte[] buf = new byte[8];
+#else
+        Span<byte> buf = stackalloc byte[8];
+#endif
+
+        Random.NextBytes(buf);
+
+        var random = BitConverter.ToInt64(buf
+#if NETSTANDARD2_0 || NET461
+            , 0);
+#else
+            );
+#endif
+
+        return new SpanId(random);
+    }
 
     /// <inheritdoc />
-    public void WriteTo(Utf8JsonWriter writer, IDiagnosticLogger? _) => writer.WriteStringValue(GetNormalizedValue());
+    public void WriteTo(Utf8JsonWriter writer, IDiagnosticLogger? _)
+    {
+        Span<byte> convertedBytes = stackalloc byte[sizeof(long)];
+        Unsafe.As<byte, long>(ref convertedBytes[0]) = _value;
+
+        // Going backwards through the array to preserve the order of the output hex string (i.e. `4e76` -> `76e4`)
+        Span<char> output = stackalloc char[16];
+        for (var i = convertedBytes.Length - 1; i >= 0; i--)
+        {
+            var value = convertedBytes[i];
+            output[(convertedBytes.Length - 1 - i) * 2] = HexChars[value >> 4];
+            output[(convertedBytes.Length - 1 - i) * 2 + 1] = HexChars[value & 0xF];
+        }
+
+        writer.WriteStringValue(output);
+    }
 
     /// <summary>
     /// Parses from string.
@@ -91,20 +115,4 @@ public readonly struct SpanId : IEquatable<SpanId>, IJsonSerializable
     /// The <see cref="Guid"/> from the <see cref="SentryId"/>.
     /// </summary>
     public static implicit operator string(SpanId id) => id.ToString();
-
-    // Note: no implicit conversion from `string` to `SpanId` as that leads to serious bugs.
-    // For example, given a method:
-    // transaction.StartChild(SpanId parentSpanId, string operation)
-    // And an *extension* method:
-    // transaction.StartChild(string operation, string description)
-    // The following code:
-    // transaction.StartChild("foo", "bar")
-    // Will resolve to the first method and not the second, which is incorrect.
-
-    /*
-    /// <summary>
-    /// A <see cref="SentryId"/> from a <see cref="Guid"/>.
-    /// </summary>
-    public static implicit operator SpanId(string value) => new(value);
-    */
 }