… CHANGELOG.md
… README.md
… Sentry-CI-Build-Linux.slnf
… Sentry-CI-Build-Windows.slnf
… Sentry-CI-Build-macOS.slnf
… Sentry.Azure.Functions.slnf
… SentryNoMobile.slnf
… SentryNoSamples.slnf
… benchmarks/Sentry.Benchmarks/LastActiveSpanBenchmarks.cs
… benchmarks/Sentry.Benchmarks/TransactionBenchmarks.cs
… samples/Sentry.Samples.Azure.Functions.Worker/host.json
… samples/Sentry.Samples.Azure.Functions.Worker/local.settings.json
… scripts/generate-solution-filters-config.yaml
… src/Sentry.Azure.Functions.Worker/.editorconfig
… test/Sentry.Azure.Functions.Worker.Tests/ApiApprovalTests.Run.DotNet6_0.verified.txt
… test/Sentry.Azure.Functions.Worker.Tests/ApiApprovalTests.Run.DotNet7_0.verified.txt
… test/Sentry.Azure.Functions.Worker.Tests/ApiApprovalTests.Run.Net4_8.verified.txt
… test/Sentry.Azure.Functions.Worker.Tests/Sentry.Azure.Functions.Worker.Tests.csproj
… test/Sentry.Azure.Functions.Worker.Tests/SentryFunctionsWorkerMiddlewareTests.cs
… test/Sentry.Testing/IsolatedRandomValuesFactory.cs
… test/Sentry.Testing/Sentry.Testing.csproj
… test/Sentry.Tests/ApiApprovalTests.Run.Core3_1.verified.txt
… test/Sentry.Tests/ApiApprovalTests.Run.DotNet6_0.verified.txt
… test/Sentry.Tests/ApiApprovalTests.Run.DotNet7_0.verified.txt
… test/Sentry.Tests/ApiApprovalTests.Run.Net4_8.verified.txt