@jamescrosswell 2023-09-20
# Important Note

I merged main into this branch and annoyingly it seems to contain a bunch of the changes from https://github.com/getsentry/sentry-dotnet/pull/2637... I'm not sure why but it makes this PR really hard to review. 

As such, I created a new branch and PR for these changes... will close this PR in favour of that branch. 

See https://github.com/getsentry/sentry-dotnet/pull/2642 instead.